+++ crates/ruma-client-api/src/sync/sync_events/v5.rs
@@ -454,7 +454,7 @@ impl Response {
 
 /// HTTP types related to a [`Response`].
 pub mod response {
-    use ruma_common::DeviceKeyAlgorithm;
+    use ruma_common::OneTimeKeyAlgorithm;
     use ruma_events::{
         receipt::SyncReceiptEvent, typing::SyncTypingEvent, AnyGlobalAccountDataEvent,
         AnyRoomAccountDataEvent, AnyToDeviceEvent,
@@ -645,15 +645,14 @@ pub mod response {
         /// For each key algorithm, the number of unclaimed one-time keys
         /// currently held on the server for a device.
         #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-        pub device_one_time_keys_count: BTreeMap<DeviceKeyAlgorithm, UInt>,
+        pub device_one_time_keys_count: BTreeMap<OneTimeKeyAlgorithm, UInt>,
 
-        /// For each key algorithm, the number of unclaimed one-time keys
-        /// currently held on the server for a device.
+        /// The unused fallback key algorithms.
         ///
         /// The presence of this field indicates that the server supports
         /// fallback keys.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub device_unused_fallback_key_types: Option<Vec<DeviceKeyAlgorithm>>,
+        pub device_unused_fallback_key_types: Option<Vec<OneTimeKeyAlgorithm>>,
     }
 
     impl E2EE {